requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
flask==3.0.0
joblib==1.3.2
python-dotenv==1.0.0
lxml==4.9.4
schedule==1.2.1
psycopg2-binary==2.9.9
